amazon-dashboard/requirements.txt
streamlit==0.81.1
pandas==0.25.1
numpy==1.17.0
plotly.express==4.14.3
wordcloud==1.8.1
matplotlib.pyplot==3.1.0
